made the GlobeSelector class out of globe-select.js
